Add sbatch script to run StrainR
Takes as input a folder containing all sequencing files for your samples of interest and runs StrainR on each sample.
You'll need to update the directory containing your files on lines 14, 18, and 19.
This script, the folder containing the database, and the StrainR scripts folder are assumed to be in the same directory. Change the paths if otherwise.
The output will appear in a folder called "results", located in the folder where your sequencing files are.